+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # Changelog
 
-Current Stable Release:    1.9.41 (April 2021)
+Current Stable Release:    1.9.45 (April 2021)
 
 -----------------------------------------------
 Release:     1.9.XX (May 2021)
 -----------------------------------------------
+- [ENH] Direct2D1: FIX namespace issues and improvements in bindings.
 
 -----------------------------------------------
-Release:     1.9.41 (April 2021)
+Release:     1.9.45 (April 2021)
 -----------------------------------------------
 KEY CHANGES:
 - [ENH] General: Update to newer SharpGen SDK and remove Vortice.Runtime.COM.